feat(restructure): Works with nvim directory - Not super robust - Add if to skip early

+++ scripts/setup_symlinks.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
 BASE_DIRECTORY="$HOME/personal_repos/my_config/"
-#FILE_SYMLINK_DIRECTORY="~/"
-DIR_SYMLINK_DIRECTORY="~/.config"
+# Add more files to symlink here.
+# Not robust enough to handle multiple branching logic depending on
+# where it should be symlinked to.
 FILES_TO_SYMLINK=(
   "dotfiles/bashrc.sh"
   "dotfiles/vimrc.vim"
@@ -11,22 +12,67 @@ FILES_TO_SYMLINK=(
 
 for filename in "${FILES_TO_SYMLINK[@]}";
 do
+
+  printf '\n%*s\n\n' "${COLUMNS:-80}" '' | tr ' ' '='
+
   filepath="${BASE_DIRECTORY}${filename}"
   printf "File to symlink: %s\n" "$filename"
   printf "Path to file: %s\n" "$filepath"
 
+  # Skip if the target does not exist
+  if [[ ! -e $filepath ]]; then
+    printf 'Skipping: %s (not found)\n' "$filepath" >&2
+    continue
+  fi
+
+  # --- Check for files ---
   if [ -f "$filepath" ];
   then
-      symlink_name=$(readlink -f "$filepath" | awk -F/ -v home="$HOME" '{sub(/\..*$/,"",$NF); print home "/." $NF}' )
+
+      symlink_name=$(
+          readlink -f "$filepath" |
+          awk -F/ -v home="$HOME" '
+            {
+              sub(/\..*$/,"",$NF)
+              print home "/." $NF
+            }'
+        )
+
       printf "Symbolic link to create: %s\n" "$symlink_name"
+
       if [ -h "$symlink_name" ];
       then
-        printf "Symbolic link exists: %s\n" "$symlink_name"
+        printf "Skipping... Symbolic link exists: %s\n" "$symlink_name"
         continue
       fi
+
       echo "Creating symlink..."
       ln -s "$filepath" "$symlink_name"
+
   fi
-  #echo "File or directory do not exist."
-  #exit 1
+
+  # --- Check for dirs ---
+  if [ -d "$filepath" ];
+  then
+      symlink_name=$(
+          readlink -f "$filepath" |
+          awk -F/ -v home="$HOME" -v config_dir="/.config/" '
+            {
+                sub(/\..*$/,"",$NF)
+                print home config_dir $NF
+            }' 
+        )
+      printf "Symbolic link to create: %s\n" "$symlink_name"
+
+      if [ -h "$symlink_name" ];
+      then
+        printf "Skipping... Symbolic link exists: %s\n" "$symlink_name"
+        continue
+      fi
+
+      echo "Creating symlink..."
+      ln -s "$filepath" "$symlink_name"
+      continue
+  fi
+
 done
